refactor(utils): combine easCredentialSupportedNetwork with blockchainCredentialSupportedNetwork
