--- libs/utils/src/getEasSupportedNetwork.ts
@@ -1,52 +1,0 @@
-import type { BlockchainNetwork } from "./types"
-
-export const easCredentialSupportedNetworks: BlockchainNetwork[] = [
-    {
-        id: "ethereum",
-        name: "Ethereum"
-    },
-    {
-        id: "sepolia",
-        name: "Ethereum (Sepolia)"
-    },
-    {
-        id: "arbitrum",
-        name: "Arbitrum"
-    },
-    {
-        id: "arbitrum-nova",
-        name: "Arbitrum Nova"
-    },
-    {
-        id: "base",
-        name: "base"
-    },
-    {
-        id: "base-sepolia",
-        name: "Base (Sepolia)"
-    },
-    {
-        id: "optimism",
-        name: "Optimism"
-    },
-    {
-        id: "optimism-sepolia-bedrock",
-        name: "Optimism (Sepolia)"
-    },
-    {
-        id: "scroll",
-        name: "scroll"
-    },
-    {
-        id: "polygon",
-        name: "Polygon"
-    },
-    {
-        id: "linea",
-        name: "Linea"
-    },
-    {
-        id: "celo",
-        name: "Celo"
-    }
-]
